Fix Streamlit Fitur

+++ db_operations.py
@@ -856,14 +856,17 @@ def delete_location(location_id: str):
     try:
         location = db.query(Location).filter(Location.location_id == location_id).first()
         if location:
+            # Check if there are units associated with this location
+            associated_units = db.query(Unit).filter(Unit.home_location == location.name).all()
+
+            if associated_units:
+                # Don't delete the location if there are associated units
+                print(f"Cannot delete location {location.name} because it has {len(associated_units)} associated unit(s)")
+                return False
+
             old_values = {'location_id': location.location_id, 'name': location.name}
 
-            # First delete all units associated with this location
-            associated_units = db.query(Unit).filter(Unit.home_location == location.name).all()
-            for unit in associated_units:
-                db.delete(unit)
-
-            # Then delete the location
+            # Delete the location (no units to worry about)
             db.delete(location)
             db.commit()
             log_audit(db, 'DELETE', 'Location', location_id, old_values=old_values)
